test(bridge): stabilize peer comparisons and split test modules
